moved error reports from root to errors module
