+++ src/app/static/src/app/root.ts
@@ -33,7 +33,6 @@ import {currentHintVersion} from "./hintVersion";
 import {ModelRunMutation, ModelRunUpdates} from "./store/modelRun/mutations";
 import {adr, ADRState, initialADRState} from "./store/adr/adr";
 import {adrUpload, ADRUploadState, initialADRUploadState} from "./store/adrUpload/adrUpload";
-import {Error} from "./generated";
 
 import {
     downloadResults,
@@ -68,9 +67,7 @@ export interface RootState extends TranslatableState {
     projects: ProjectsState
     currentUser: string,
     downloadResults: DownloadResultsState,
-    updatingLanguage: boolean,
-    errorReportError: Error | null
-    errorReportSuccess: boolean
+    updatingLanguage: boolean
 }
 
 export interface ReadyState {
@@ -139,8 +136,6 @@ export const emptyState = (): RootState => {
         language: Language.en,
         version: currentHintVersion,
         updatingLanguage: false,
-        errorReportError: null,
-        errorReportSuccess: false,
         hintrVersion: initialHintrVersionState(),
         adr: initialADRState(),
         genericChart: initialGenericChartState(),
