+++ src/app/static/src/app/store/errors/errors.ts
@@ -2,14 +2,19 @@ import {Module} from "vuex";
 import {RootState} from "../../root";
 import {Error} from "../../generated";
 import {mutations} from "./mutations";
+import {actions} from "./actions";
 
 export interface ErrorsState {
-    errors: Error[]
+    errors: Error[],
+    errorReportError: Error | null
+    errorReportSuccess: boolean
 }
 
 export const initialErrorsState = (): ErrorsState => {
     return {
-        errors: []
+        errors: [],
+        errorReportError: null,
+        errorReportSuccess: false,
     }
 };
 
@@ -18,5 +23,6 @@ const namespaced = true;
 export const errors: Module<ErrorsState, RootState> = {
     namespaced,
     state: initialErrorsState(),
-    mutations
+    mutations,
+    actions
 };
