+++ src/app/static/src/app/store/root/mutations.ts
@@ -17,7 +17,6 @@ import {initialModelCalibrateState} from "../modelCalibrate/modelCalibrate";
 import {initialADRUploadState} from "../adrUpload/adrUpload";
 import {initialDownloadResultsState} from "../downloadResults/downloadResults";
 import {initialGenericChartState} from "../genericChart/genericChart";
-import {Error} from "../../generated";
 
 export enum RootMutation {
     Reset = "Reset",
@@ -26,9 +25,7 @@ export enum RootMutation {
     ResetOutputs = "ResetOutputs",
     SetProject = "SetProject",
     ResetDownload = "ResetDownload",
-    SetUpdatingLanguage = "SetUpdatingLanguage",
-    ErrorReportError = "ErrorReportError",
-    ErrorReportSuccess = "ErrorReportSuccess"
+    SetUpdatingLanguage = "SetUpdatingLanguage"
 }
 
 export const mutations: MutationTree<RootState> = {
@@ -41,8 +38,6 @@ export const mutations: MutationTree<RootState> = {
         const resetState: RootState = {
             version: state.version,
             hintrVersion: state.hintrVersion,
-            errorReportError: null,
-            errorReportSuccess: false,
             language: state.language,
             updatingLanguage: false,
             adr: state.adr,
@@ -160,17 +155,6 @@ export const mutations: MutationTree<RootState> = {
     [RootMutation.SetUpdatingLanguage](state: RootState, action: PayloadWithType<boolean>) {
         state.updatingLanguage = action.payload;
     },
-
-    [RootMutation.ErrorReportError](state: RootState, action: PayloadWithType<Error>) {
-        state.errorReportError = action.payload;
-        state.errorReportSuccess = false;
-    },
-
-    [RootMutation.ErrorReportSuccess](state: RootState) {
-        state.errorReportSuccess = true;
-        state.errorReportError = null
-    },
-
     ...languageMutations
 
 };
